+++ src/applications/financial-status-report/config/form.js
@@ -424,6 +424,15 @@ const formConfig = {
           uiSchema: pages.bankruptcyHistory.uiSchema,
           schema: pages.bankruptcyHistory.schema,
         },
+        bankruptcyHistoryRecords: {
+          path: 'bankruptcy-history-records',
+          title: 'Bankruptcy history',
+          uiSchema: pages.bankruptcyHistoryRecords.uiSchema,
+          schema: pages.bankruptcyHistoryRecords.schema,
+          depends: formData =>
+            formData.bankruptcyHistory.hasBeenAdjudicated ===
+            'Yes, I have been adjudicated as bankrupt.',
+        },
       },
     },
   },
